+++ packages-content-model/roosterjs-content-model-editor/lib/editor/createContentModelEditorCore.ts
@@ -1,6 +1,8 @@
 import ContentModelCopyPastePlugin from './corePlugins/ContentModelCopyPastePlugin';
 import ContentModelTypeInContainerPlugin from './corePlugins/ContentModelTypeInContainerPlugin';
+import { ContentModelEditorCore } from '../publicTypes/ContentModelEditorCore';
 import { ContentModelEditorOptions } from '../publicTypes/IContentModelEditor';
+import { ContentModelPluginState } from '../publicTypes/pluginState/ContentModelPluginState';
 import { ContentModelSegmentFormat } from 'roosterjs-content-model-types';
 import { CoreCreator, EditorCore, ExperimentalFeatures } from 'roosterjs-editor-types';
 import { createContentModel } from './coreApi/createContentModel';
@@ -13,10 +15,6 @@ import { getSelectionRangeEx } from './coreApi/getSelectionRangeEx';
 import { setContentModel } from './coreApi/setContentModel';
 import { switchShadowEdit } from './coreApi/switchShadowEdit';
 import { tablePreProcessor } from './overrides/tablePreProcessor';
-import {
-    ContentModelEditorCore,
-    ContentModelPluginState,
-} from '../publicTypes/ContentModelEditorCore';
 
 /**
  * Editor Core creator for Content Model editor
