+++ packages-content-model/roosterjs-content-model-editor/lib/index.ts
@@ -7,7 +7,6 @@ export {
     ContentModelEditorCore,
     CreateContentModel,
     SetContentModel,
-    ContentModelPluginState,
 } from './publicTypes/ContentModelEditorCore';
 export {
     default as ContentModelBeforePasteEvent,
@@ -81,10 +80,7 @@ export { default as keyboardDelete } from './publicApi/editing/keyboardDelete';
 export { default as ContentModelEditor } from './editor/ContentModelEditor';
 export { default as isContentModelEditor } from './editor/isContentModelEditor';
 export { default as ContentModelFormatPlugin } from './editor/plugins/ContentModelFormatPlugin';
-export {
-    default as ContentModelEditPlugin,
-    ContentModelEditPluginState,
-} from './editor/plugins/ContentModelEditPlugin';
+export { default as ContentModelEditPlugin } from './editor/plugins/ContentModelEditPlugin';
 export { default as ContentModelPastePlugin } from './editor/plugins/PastePlugin/ContentModelPastePlugin';
 export { default as ContentModelTypeInContainerPlugin } from './editor/corePlugins/ContentModelTypeInContainerPlugin';
 export { default as ContentModelCopyPastePlugin } from './editor/corePlugins/ContentModelCopyPastePlugin';
@@ -96,3 +92,6 @@ export { combineBorderValue, extractBorderValues } from './domUtils/borderValues
 export { updateImageMetadata } from './domUtils/metadata/updateImageMetadata';
 export { updateTableCellMetadata } from './domUtils/metadata/updateTableCellMetadata';
 export { updateTableMetadata } from './domUtils/metadata/updateTableMetadata';
+
+export { ContentModelEditPluginState } from './publicTypes/pluginState/ContentModelEditPluginState';
+export { ContentModelPluginState } from './publicTypes/pluginState/ContentModelPluginState';
